# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for SlideShare - real-time presentation sharing
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PresentationStatus = typeof PresentationStatus[keyof typeof PresentationStatus];


export const PresentationStatus = {
  pending: 'pending',
  converting: 'converting',
  ready: 'ready',
  error: 'error',
} as const;

export interface Presentation {
  id: string;
  title: string;
  fileName: string;
  fileObjectPath: string;
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  slideCount?: number | null;
  currentSlide: number;
  maxRevealedSlide: number;
  status: PresentationStatus;
  isPublished: boolean;
  isLive: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface PresentationInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  fileObjectPath: string;
  /** @minLength 1 */
  contentType: string;
}

export interface PresentationUpdate {
  password: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
}

export interface AdminPassword {
  password: string;
}

export interface DeletePresentationResult {
  success: boolean;
}

export interface PublishPresentation {
  password: string;
  published: boolean;
}

export interface LivePresentationUpdate {
  password: string;
  live: boolean;
}

export interface QuizQuestion {
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
}

export type QuizStatus = typeof QuizStatus[keyof typeof QuizStatus];


export const QuizStatus = {
  ready: 'ready',
} as const;

export interface Quiz {
  id: string;
  presentationId: string;
  status: QuizStatus;
  /**
     * @minItems 20
     * @maxItems 20
     */
  questions: QuizQuestion[];
  totalQuestions: number;
  createdAt: string;
}

export interface QuizAttemptInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  studentName: string;
  /**
     * @minItems 20
     * @maxItems 20
     * @items.minimum 0
     * @items.maximum 3
     */
  answers: number[];
}

export interface QuizAnswers {
  /**
     * @minItems 20
     * @maxItems 20
     * @items.minimum 0
     * @items.maximum 3
     */
  answers: number[];
}

export interface QuizAttempt {
  id: string;
  quizId: string;
  studentName: string;
  score: number;
  totalQuestions: number;
  createdAt: string;
}

export interface QuizReviewItem {
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
  /**
     * @minimum 0
     * @maximum 3
     */
  selectedIndex: number;
  /**
     * @minimum 0
     * @maximum 3
     */
  correctIndex: number;
  isCorrect: boolean;
}

export interface QuizSubmissionResult {
  id: string;
  quizId: string;
  studentName: string;
  score: number;
  totalQuestions: number;
  createdAt: string;
  /**
     * @minItems 20
     * @maxItems 20
     */
  review: QuizReviewItem[];
}

export interface QuizGradeResult {
  score: number;
  totalQuestions: number;
  /**
     * @minItems 20
     * @maxItems 20
     */
  review: QuizReviewItem[];
}

export interface SlideUpdate {
  password: string;
  /** @minimum 0 */
  slideIndex: number;
}

export interface ConversionResult {
  status: string;
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  slideCount?: number | null;
  /** @nullable */
  error?: string | null;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}