# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SlideShare - real-time presentation sharing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /presentations:
    get:
      operationId: listPresentations
      tags: [presentations]
      summary: List all presentations
      responses:
        "200":
          description: List of presentations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Presentation"
    post:
      operationId: createPresentation
      tags: [presentations]
      summary: Create a new presentation record after upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresentationInput"
      responses:
        "201":
          description: Created presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presentation"
        "400":
          description: Invalid input
  /presentations/{id}:
    get:
      operationId: getPresentation
      tags: [presentations]
      summary: Get presentation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presentation"
        "404":
          description: Not found
    patch:
      operationId: updatePresentation
      tags: [presentations]
      summary: Rename a presentation (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresentationUpdate"
      responses:
        "200":
          description: Renamed presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presentation"
        "400":
          description: Invalid title
        "401":
          description: Unauthorized - wrong password
        "404":
          description: Not found
    delete:
      operationId: deletePresentation
      tags: [presentations]
      summary: Delete a presentation and its uploaded file (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPassword"
      responses:
        "200":
          description: Deleted presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletePresentationResult"
        "401":
          description: Unauthorized - wrong password
        "404":
          description: Not found
  /presentations/{id}/slide:
    patch:
      operationId: updateSlide
      tags: [presentations]
      summary: Update the current slide (presenter only, requires password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlideUpdate"
      responses:
        "200":
          description: Updated presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presentation"
        "401":
          description: Unauthorized - wrong password
        "404":
          description: Not found
  /presentations/{id}/publish:
    patch:
      operationId: publishPresentation
      tags: [presentations]
      summary: Publish or unpublish a presentation (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishPresentation"
      responses:
        "200":
          description: Updated presentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presentation"
        "401":
          description: Unauthorized - wrong password
        "404":
          description: Not found
...
  /presentations/{id}/convert:
    post:
      operationId: convertPresentation
      tags: [presentations]
      summary: Trigger PPT to PDF conversion for a presentation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResult"
        "404":
          description: Not found
...
components:
  schemas:
...
    Presentation:
      type: object
      required: [id, title, fileName, fileObjectPath, status, currentSlide, maxRevealedSlide, isPublished, isLive, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        fileName:
          type: string
        fileObjectPath:
          type: string
        pdfObjectPath:
          type: ["string", "null"]
        slideCount:
          type: ["number", "null"]
        currentSlide:
          type: number
        maxRevealedSlide:
          type: number
        status:
          type: string
          enum: [pending, converting, ready, error]
        isPublished:
          type: boolean
        isLive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PresentationInput:
      type: object
      required: [title, fileName, fileObjectPath, contentType]
      properties:
        title:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        fileObjectPath:
          type: string
          minLength: 1
        contentType:
          type: string
          minLength: 1
    PresentationUpdate:
      type: object
      required: [password, title]
      properties:
        password:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 200
    AdminPassword:
      type: object
      required: [password]
      properties:
        password:
          type: string
    DeletePresentationResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    PublishPresentation:
      type: object
      required: [password, published]
      properties:
        password:
          type: string
        published:
          type: boolean
...
    SlideUpdate:
      type: object
      required: [password, slideIndex]
      properties:
        password:
          type: string
        slideIndex:
          type: number
          minimum: 0
    ConversionResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
        pdfObjectPath:
          type: ["string", "null"]
        slideCount:
          type: ["number", "null"]
        error:
          type: ["string", "null"]
...