# lib/api-zod/src/generated/types/conversionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for SlideShare - real-time presentation sharing
 * OpenAPI spec version: 0.1.0
 */

export interface ConversionResult {
  status: string;
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  slideCount?: number | null;
  /** @nullable */
  error?: string | null;
}