# lib/api-zod/src/generated/types/presentation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for SlideShare - real-time presentation sharing
 * OpenAPI spec version: 0.1.0
 */
import type { PresentationStatus } from './presentationStatus';

export interface Presentation {
  id: string;
  title: string;
  fileName: string;
  fileObjectPath: string;
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  slideCount?: number | null;
  currentSlide: number;
  maxRevealedSlide: number;
  status: PresentationStatus;
  isPublished: boolean;
  isLive: boolean;
  createdAt: Date;
  updatedAt?: Date;
}