# lib/api-zod/src/generated/types/slideUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for SlideShare - real-time presentation sharing
 * OpenAPI spec version: 0.1.0
 */

export interface SlideUpdate {
  password: string;
  /** @minimum 0 */
  slideIndex: number;
}